mytab_indp_K_Populations {myfunR}	R Documentation

mytab_indp_K_Populations

Description

It gives a output table for testing the significant difference between the means of k number of independent populations.

Usage


mytab_indp_K_Populations(x, group, xname = deparse(substitute(x)))

Arguments

x
dependent numeric variable

group
independent categorical variable

xname
name of the dependent variable

Value

Returns a data frame containing the descriptive stat along with the 1 way ANOVA F test stat., 1 way ANOVA P Value & Non parametric(Kruskal Wallis test) P Value.

[Package myfunR version 0.1.0 Index]