mytab_paired_one_Populations {myfunR}	R Documentation

mytab_paired_one_Populations

Description

It gives a output table for testing the significant difference between the pre and post means of a independent populations.

Usage


mytab_paired_one_Populations(x.before, x.after, var)

Arguments

x.before
numeric vector for the pre score.

x.after
numeric vector for the post score.

var
name of the variable/ population.

Value

Returns a data frame containing the descriptive stat along with the Paired.t- test test statistic, Confidence Interval, Paired.t- test P-value & Non parametric(Wilcoxon Signed Rank test) P Value.

[Package myfunR version 0.1.0 Index]