mycenter {myfunR}	R Documentation

mycenter

Description

it helps to fix the table's heading into the center.

Usage


mycenter(text, lsp, usp)

Arguments

text
Table's heading

lsp
no. of spaces/blanks spaces in the left side of the text.

usp
no. of spaces/blanks spaces in the right side of the text.

Value

character

Examples


mycat(mycenter("This is my Table"),20,0)

[Package myfunR version 0.1.0 Index]