mytab_indep_two_Populations {myfunR}	R Documentation

mytab_indep_two_Populations

Description

It gives a output table for testing the significant difference between the means of two independent populations.

Usage


mytab_indep_two_Populations(x, group, xname = deparse(substitute(x)))

Arguments

x
input dependent numeric variable.

group
independent character vector for the names of the two populations.

xname
name of the input dependent numeric variable.

Value

Returns a data frame containing the descriptive stat along with the indep. t-test stat., indep. t-test P Value & Non parametric(Mann Whitney U test) P Value.

[Package myfunR version 0.1.0 Index]