mytab_two_two_Contingency {myfunR}	R Documentation

mytab_two_two_Contingency

Description

It gives an output table for testing the significant difference between the probabilities of two independent populations,i.e., for 2x2 contingency table.

Usage


mytab_two_two_Contingency(x, y, xlev = deparse(substitute(x)))

Arguments

x
indep. character vector.

y
dep. character vector.

xlev
levels of x.

Value

Returns a data frame containing the probs., percentages along with the Chi-square Test P-Value & Fisher's Exact Test P-Value for Non-Parametric.

[Package myfunR version 0.1.0 Index]